+++ sdk/resourcemanager/azure-resourcemanager-redis/src/main/java/com/azure/resourcemanager/redis/models/RedisCache.java
@@ -8,6 +8,7 @@ import com.azure.resourcemanager.redis.RedisManager;
 import com.azure.resourcemanager.redis.fluent.models.RedisResourceInner;
 import com.azure.resourcemanager.resources.fluentcore.arm.models.GroupableResource;
 import com.azure.resourcemanager.resources.fluentcore.arm.models.HasId;
+import com.azure.resourcemanager.resources.fluentcore.arm.models.Resource;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsListingPrivateEndpointConnection;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsListingPrivateLinkResource;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsUpdatingPrivateEndpointConnection;
@@ -279,11 +280,11 @@ public interface RedisCache
              * @param redisVersion the redisVersion value to set.
              * @return the next stage of Redis Cache with Premium SKU definition.
              */
-            WithCreate withRedisVersion(MajorVersion redisVersion);
+            WithCreate withRedisVersion(RedisVersion redisVersion);
         }
 
         /** A Redis Cache definition with Premium Sku specific functionality. */
-        interface WithPremiumSkuCreate extends WithCreate {
+        interface WithPremiumSkuCreate extends DefinitionStages.WithCreate {
 
             /**
              * The number of shards to be created on a Premium Cluster Cache.
@@ -431,7 +432,7 @@ public interface RedisCache
     /** The template for a Redis Cache update operation, containing all the settings that can be modified. */
     interface Update
         extends Appliable<RedisCache>,
-            UpdateWithTags<Update>,
+            Resource.UpdateWithTags<Update>,
             UpdateStages.WithSku,
             UpdateStages.WithNonSslPort,
             UpdateStages.WithRedisConfiguration {
@@ -532,7 +533,7 @@ public interface RedisCache
          * @param redisVersion the redisVersion value to set
          * @return the next stage of Redis Cache definition.
          */
-        Update withRedisVersion(MajorVersion redisVersion);
+        Update withRedisVersion(RedisVersion redisVersion);
     }
 
     /**
@@ -540,13 +541,19 @@ public interface RedisCache
      * for version 4.x.x, use V4
      * for version 6.x.x, use V6
      */
-    enum  MajorVersion {
-        V6("6"), // version 6.x.x
-        V4("4"); // version 4.x.x
+    enum RedisVersion {
+        /**
+         * version 6.x.x
+         */
+        V6("6"),
+        /**
+         * version 4.x.x
+         */
+        V4("4");
 
         private final String value;
 
-        MajorVersion(String value) {
+        RedisVersion(String value) {
             this.value = value;
         }