+++ sdk/resourcemanager/azure-resourcemanager-redis/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 2.10.0-beta.1 (Unreleased)
 
 ### Features Added
-mgmt, support redisVersion in convenient layer of redis
+
+ - supported `redisVersion` in convenient layer of `redis`
 
 ## 2.9.0 (2021-10-21)